# Conflicts:
#	pom.xml
#	src/main/java/com/qf/ecms/domain/entity/ColorSize.java
#	src/main/java/com/qf/ecms/domain/entity/ShopCart.java
#	src/main/java/com/qf/ecms/domain/entity/User.java
#	src/main/java/com/qf/ecms/mapper/ColorSizeMapper.java
#	src/main/java/com/qf/ecms/utils/ErrorStatus.java
#	src/main/resources/mapper/ColorSizeMapper.xml
#	src/main/resources/spring-mybatis.xml

+++ src/main/java/com/qf/ecms/domain/entity/ShopCart.java
@@ -1,61 +1,62 @@
 package com.qf.ecms.domain.entity;
 
+import java.math.BigDecimal;
 import lombok.Data;
 
 @Data
 public class ShopCart {
     /**
-     * 购物车id
-     */
+    * 购物车id
+    */
     private Integer cartId;
 
     /**
-     * 用户id
-     */
+    * 用户id
+    */
     private Integer userId;
 
     /**
-     * 购买的商品数量
-     */
+    * 购买的商品数量
+    */
     private Integer count;
 
     /**
-     * 商品id
-     */
+    * 商品id
+    */
     private Integer cid;
 
     /**
-     * 商品标题
-     */
+    * 商品标题
+    */
     private String commodityTitle;
 
     /**
-     * 商品封面图
-     */
+    * 商品封面图
+    */
     private String commentPicture;
 
     /**
-     * 颜色名
-     */
+    * 颜色名
+    */
     private String colorName;
 
     /**
-     * 尺寸名
-     */
+    * 尺寸名
+    */
     private String sizeName;
 
     /**
-     * 颜色尺寸表主键
-     */
+    * 颜色尺寸表主键
+    */
     private Integer csid;
 
     /**
-     * 现价
-     */
-    private Long presentPrice;
+    * 现价
+    */
+    private BigDecimal presentPrice;
 
     /**
-     * 状态码 1:正常 0:删除
-     */
+    * 状态码 1:正常 0:删除
+    */
     private Integer status;
 }
